+++ app/src/main/java/app/revanced/manager/util/permission/NotificationPermission.kt
@@ -4,13 +4,17 @@ import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
+import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 
 fun Context.hasNotificationPermission(): Boolean {
-    return if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
-        true
-    } else {
-        ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS) ==
-            PackageManager.PERMISSION_GRANTED
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        val hasPermission =
+            ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS) ==
+                PackageManager.PERMISSION_GRANTED
+        if (!hasPermission) {
+            return false
+        }
     }
+    return NotificationManagerCompat.from(this).areNotificationsEnabled()
 }